clean code teamcard component
